Make OutputStream constructor not throw
Moved the header-writing into the different writing functions,
tracking whether it had been done with a state variable. This allows
the stream to be constructed in functions that are not expected to
throw with the write operations called in other functions that ARE
expected to throw.

I also added a variant constructor that takes a deflater, intending to
call it from the hadoop methods that take a Compressor. It turned out
not to be necessary, but it is harder to take out and the code is
possibly a little cleaner with it in there.